chg: [migration] head and verif

+++ migrations/versions/598df1c175e3_add_auth_provider_to_user.py
@@ -1,7 +1,7 @@
 """add auth_provider to user
 
 Revision ID: 598df1c175e3
-Revises: 26c7e5d76677
+Revises: 697786a00ccd
 Create Date: 2026-02-20 21:54:31.776484
 
 """
@@ -11,16 +11,26 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision = '598df1c175e3'
-down_revision = '26c7e5d76677'
+down_revision = '697786a00ccd'
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
+    bind = op.get_bind()
+    inspector = sa.inspect(bind)
+    existing_columns = [col['name'] for col in inspector.get_columns('user')]
+
     with op.batch_alter_table('user', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('auth_provider', sa.String(length=32), server_default='local', nullable=False))
+        if 'auth_provider' not in existing_columns:
+            batch_op.add_column(sa.Column('auth_provider', sa.String(length=32), server_default='local', nullable=False))
 
 
 def downgrade():
+    bind = op.get_bind()
+    inspector = sa.inspect(bind)
+    existing_columns = [col['name'] for col in inspector.get_columns('user')]
+
     with op.batch_alter_table('user', schema=None) as batch_op:
-        batch_op.drop_column('auth_provider')
+        if 'auth_provider' in existing_columns:
+            batch_op.drop_column('auth_provider')
